refactor(api): разрезать system_tools.go по доменам и покрыть хендлер тестами
Файл на 1386 строк держал пять доменов, разложенных в internal/sys по
отдельным пакетам. Разрез механический, аннотации swagger переехали вместе
с функциями (32 @Router на месте, дрейфа нет):

- system_tools.go   85 — хендлер, конструктор, expert-гейт, события
- system_files.go  431 — файловый менеджер
- system_scripts.go 309 — статус и запуск скриптов
- system_opkg.go   243 — пакеты
- system_services.go 186 — службы init.d
- system_ports.go  119 — сокеты
- system_procmon.go 81 — монитор процессов

Тесты на слой хендлера (раньше их не было ни одного): гейт usage level в
обе стороны, чтение/запись внутри песочницы, отказ на путь снаружи, отказ
на запись в read-only корень, запуск вне корней, отсутствие args у
script-action, флаг вместо имени пакета в opkg, kill PID 1, метод-гейт,
режим chmod.

Проверено мутациями: снятие expert-гейта, чтение мимо Resolve и старая
обрезка ведущего нуля в chmod валят соответствующие тесты.
